Split application.js into core.js and application.js (hitobito/hitobito_sac_cas#2538)
When adding a new layout we may not want/need every JS dependency from application.js, the idea is to have a leaner core pack for standard utilties we actually want everywhere (polyfills, bootstrap, stimulus and so on).

application.js was something above 4MB, we split this up to now core (2.9MB) and load the rest for application only. We could actually decrease this more when our custom scripts from core we still load in core.js into generic and feature specific scripts. Same goes for stimulus controllers
