+++ app/views/layouts/oauth.html.haml
@@ -13,7 +13,7 @@
     = stylesheet_pack_tag 'oauth', media: 'screen', 'data-turbo-track': true
     - if Rails.env.test?
       = stylesheet_pack_tag 'disable_animations'
-    = javascript_pack_tag 'application', 'data-turbo-track': true
+    = javascript_pack_tag 'core', 'data-turbo-track': true
 
     = yield(:head)
