+++ spec/features/stimulus/autosubmit_controller_spec.rb
@@ -19,7 +19,7 @@ describe "Autosubmit Stimulus Controller", js: true do
         render inline: <<~HTML
           <head>
             #{stylesheet_pack_tag "application", media: "screen", "data-turbo-track": true}
-            #{javascript_pack_tag "application", "data-turbo-track": true}
+            #{javascript_pack_tag "core", "data-turbo-track": true}
             <link rel="icon" href="data:image/x-icon;," type="image/x-icon">
           </head>
           <body>
@@ -33,7 +33,7 @@ describe "Autosubmit Stimulus Controller", js: true do
         render inline: <<~HTML
           <head>
             #{stylesheet_pack_tag "application", media: "screen", "data-turbo-track": true}
-            #{javascript_pack_tag "application", "data-turbo-track": true}
+            #{javascript_pack_tag "core", "data-turbo-track": true}
             <link rel="icon" href="data:image/x-icon;," type="image/x-icon">
           </head>
           <body>
